RATEST-286: Refactor allergy workflow
