--- qaframework-bdd-tests/src/test/resources/features/refapp-2.x/10-allergy/allergy.feature
@@ -1,12 +1,0 @@
-Feature: Allery Management
-
-  Background:
-  Given a user clicks on Allergies link from the Patient dashboard
-  Then the system loads the Allergy page
-
-  @selenium
-  @allergy
-  Scenario: Allergy management
-   When a user adds a known allergy into the system
-   Then a user edits a known allergy
-   And a user deletes a known allergy
